get_percent_var_v3.R
the code now makes a stacked bar plot, but the formatting is off in terms of the assigning a name to each stacked bar as the filenames are way too long, i'm pretty sure there is a way to get the names to be vertically alligned but haven't been able to get it going. ggplot seemed to be having a hard time handling the dataframe from my side, you might know better.